+++ samples/Sample_05_Chart.php
@@ -209,6 +209,7 @@ function fnSlide_BarStacked(PhpPresentation $objPHPPresentation) {
     $shape->getLegend()->getBorder()->setLineStyle( Border::LINE_SINGLE );
     $shape->getLegend()->getFont()->setItalic( true );
 }
+
 function fnSlide_BarPercentStacked(PhpPresentation $objPHPPresentation) {
     global $oFill;
     global $oShadow;
@@ -492,17 +493,19 @@ function fnSlide_Scatter(PhpPresentation $objPHPPresentation) {
     $lineChart = new Scatter();
     $series = new Series('Downloads', $seriesData);
     $series->setShowSeriesName(true);
+    $series->getMarker()->setSymbol(\PhpOffice\PhpPresentation\Shape\Chart\Marker::SYMBOL_DASH);
+    $series->getMarker()->setSize(10);
     $lineChart->addSeries($series);
     
     // Create a shape (chart)
     echo date('H:i:s') . ' Create a shape (chart)'.EOL;
     $shape = $currentSlide->createChartShape();
     $shape->setName('PHPPresentation Daily Download Distribution')
-    ->setResizeProportional(false)
-    ->setHeight(550)
-    ->setWidth(700)
-    ->setOffsetX(120)
-    ->setOffsetY(80);
+        ->setResizeProportional(false)
+        ->setHeight(550)
+        ->setWidth(700)
+        ->setOffsetX(120)
+        ->setOffsetY(80);
     $shape->setShadow($oShadow);
     $shape->setFill($oFill);
     $shape->getBorder()->setLineStyle(Border::LINE_SINGLE);