+++ samples/Sample_05_Chart_Line.php
@@ -20,7 +20,7 @@ $objPHPPresentation = new PhpPresentation();
 
 // Set properties
 echo date('H:i:s') . ' Set properties' . EOL;
-$objPHPPresentation->getProperties()->setCreator('PHPOffice')->setLastModifiedBy('PHPPresentation Team')->setTitle('Sample 07 Title')->setSubject('Sample 07 Subject')->setDescription('Sample 07 Description')->setKeywords('office 2007 openxml libreoffice odt php')->setCategory('Sample Category');
+$objPHPPresentation->getDocumentProperties()->setCreator('PHPOffice')->setLastModifiedBy('PHPPresentation Team')->setTitle('Sample 07 Title')->setSubject('Sample 07 Subject')->setDescription('Sample 07 Description')->setKeywords('office 2007 openxml libreoffice odt php')->setCategory('Sample Category');
 
 // Remove first slide
 echo date('H:i:s') . ' Remove first slide' . EOL;
@@ -36,13 +36,13 @@ $oShadow->setVisible(true)->setDirection(45)->setDistance(10);
 // Generate sample data for chart
 echo date('H:i:s') . ' Generate sample data for chart' . EOL;
 $seriesData = array(
-	'Monday' => 12,
-	'Tuesday' => 15,
-	'Wednesday' => 13,
-	'Thursday' => 17,
-	'Friday' => 14,
-	'Saturday' => 9,
-	'Sunday' => 7
+    'Monday' => 12,
+    'Tuesday' => 15,
+    'Wednesday' => 13,
+    'Thursday' => 17,
+    'Friday' => 14,
+    'Saturday' => 9,
+    'Sunday' => 7
 );
 
 // Create templated slide
@@ -79,6 +79,10 @@ $currentSlide = createTemplatedSlide($objPHPPresentation);
 // Create a line chart (that should be inserted in a shape)
 echo date('H:i:s') . ' Create a line chart (that should be inserted in a chart shape)' . EOL;
 $lineChart1 = clone $lineChart;
+$series1 = $lineChart1->getSeries();
+$series1[0]->getMarker()->setSymbol(\PhpOffice\PhpPresentation\Shape\Chart\Marker::SYMBOL_DIAMOND);
+$series1[0]->getMarker()->setSize(7);
+$lineChart1->setSeries($series1);
 
 // Create a shape (chart)
 echo date('H:i:s') . ' Create a shape (chart)' . EOL;
@@ -98,9 +102,11 @@ $currentSlide = createTemplatedSlide($objPHPPresentation);
 // Create a line chart (that should be inserted in a shape)
 echo date('H:i:s') . ' Create a line chart (that should be inserted in a chart shape)' . EOL;
 $lineChart2 = clone $lineChart;
-$series2 = $lineChart2->getData();
+$series2 = $lineChart2->getSeries();
 $series2[0]->getFill()->setFillType(Fill::FILL_SOLID);
-$lineChart2->setData($series2);
+$series2[0]->getMarker()->setSymbol(\PhpOffice\PhpPresentation\Shape\Chart\Marker::SYMBOL_TRIANGLE);
+$series2[0]->getMarker()->setSize(10);
+$lineChart2->setSeries($series2);
 
 // Create a shape (chart)
 echo date('H:i:s') . ' Create a shape (chart)' . EOL;
@@ -117,5 +123,5 @@ $currentSlide->addShape($shape2);
 echo EOL . write($objPHPPresentation, basename(__FILE__, '.php'), $writers);
 
 if (!CLI) {
-	include_once 'Sample_Footer.php';
+    include_once 'Sample_Footer.php';
 }