+++ tests/PhpPresentation/Tests/Shape/Chart/SeriesTest.php
@@ -17,6 +17,7 @@
 
 namespace PhpOffice\PhpPresentation\Tests\Shape\Chart;
 
+use PhpOffice\PhpPresentation\Shape\Chart\Marker;
 use PhpOffice\PhpPresentation\Shape\Chart\Series;
 use PhpOffice\PhpPresentation\Style\Fill;
 use PhpOffice\PhpPresentation\Style\Font;
@@ -39,6 +40,7 @@ class SeriesTest extends \PHPUnit_Framework_TestCase
         $this->assertEquals('Series Title', $object->getTitle());
         $this->assertInternalType('array', $object->getValues());
         $this->assertEmpty($object->getValues());
+        $this->assertInstanceOf('PhpOffice\\PhpPresentation\\Shape\\Chart\\Marker', $object->getMarker());
     }
     
     public function testDataLabelNumFormat()
@@ -113,6 +115,14 @@ class SeriesTest extends \PHPUnit_Framework_TestCase
         $this->assertEmpty($object->getHashIndex());
         $this->assertInstanceOf('PhpOffice\\PhpPresentation\\Shape\\Chart\\Series', $object->setLabelPosition(Series::LABEL_INSIDEBASE));
         $this->assertEquals(Series::LABEL_INSIDEBASE, $object->getLabelPosition());
+    }
+
+    public function testMarker()
+    {
+        $object = new Series();
+
+        $this->assertInstanceOf('PhpOffice\\PhpPresentation\\Shape\\Chart\\Series', $object->setMarker(new Marker()));
+        $this->assertInstanceOf('PhpOffice\\PhpPresentation\\Shape\\Chart\\Marker', $object->getMarker());
     }
 
     public function testShowCategoryName()